backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-dotenv==1.0.0
pydantic==2.5.0
sqlalchemy==2.0.23
alembic==1.13.0
psycopg2-binary==2.9.9
chromadb==0.4.18
sentence-transformers==2.2.2
transformers==4.36.0
torch==2.1.1
numpy==1.24.3
pandas==2.1.4
PyPDF2==3.0.1
python-docx==1.1.0
beautifulsoup4==4.12.2
requests==2.31.0
openai==1.3.7
anthropic==0.7.8
langchain==0.0.350
langchain-community==0.0.5
langchain-openai==0.0.2
tiktoken==0.5.2
redis==5.0.1
celery==5.3.4
httpx==0.25.2
aiofiles==23.2.1
Pillow==10.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-magic==0.4.27
validators==0.22.0